requirements.txt
pyspark>=3.4
dagster>=1.7
dagster-webserver>=1.7
dagster-pyspark>=0.22
pydantic>=2.7
pytest>=8.0
pyyaml>=6.0
